🧹 Move Configuration to a better place
